make the check* signatures more uniform, all take headers, mmh. checkPublicKey is still an outlier because it needs access both to headers and body
