+++ asserts/account_key.go
@@ -48,20 +48,6 @@ func (ak *AccountKey) Until() time.Time {
 	return ak.until
 }
 
-// TODO: move check* helpers to separate file if they get reused
-
-func checkRFC3339Date(ab *AssertionBase, name string) (time.Time, error) {
-	dateStr := ab.Header(name)
-	if dateStr == "" {
-		return time.Time{}, fmt.Errorf("%v header is mandatory", name)
-	}
-	date, err := time.Parse(time.RFC3339, dateStr)
-	if err != nil {
-		return time.Time{}, fmt.Errorf("%v header is not a RFC3339 date: %v", name, err)
-	}
-	return date, nil
-}
-
 func checkPublicKey(ab *AssertionBase, fingerprintName string) (PublicKey, error) {
 	pubKey, err := parsePublicKey(ab.Body())
 	if err != nil {
@@ -78,14 +64,15 @@ func checkPublicKey(ab *AssertionBase, fingerprintName string) (PublicKey, error
 }
 
 func buildAccountKey(assert AssertionBase) (Assertion, error) {
-	if assert.Header("account-id") == "" {
-		return nil, fmt.Errorf("account-id header is mandatory")
-	}
-	since, err := checkRFC3339Date(&assert, "since")
+	_, err := checkMandatory(assert.headers, "account-id")
 	if err != nil {
 		return nil, err
 	}
-	until, err := checkRFC3339Date(&assert, "until")
+	since, err := checkRFC3339Date(assert.headers, "since")
+	if err != nil {
+		return nil, err
+	}
+	until, err := checkRFC3339Date(assert.headers, "until")
 	if err != nil {
 		return nil, err
 	}
